Add files via upload
I am uploading this now, but I haven't made any changes to these files except their names since June this year. This is my initial approach to creating this calculator, but I hadn't laid out specific details or anything I wanted to accomplish with this, so the code is a bit all over the place.